+++ customxepr/config/main.py
@@ -29,7 +29,7 @@ DEFAULTS = [
               'y': 0,
               'width': 1024,
               'height': 650,
-              'splitter': [824, 200],
+              'auto_plot_results': True
               }),
             ('CustomXepr',
              {
@@ -51,7 +51,7 @@ DEFAULTS = [
 #    or if you want to *rename* options, then you need to do a MAJOR update in
 #    version, e.g. from 3.0.0 to 4.0.0
 # 3. You don't need to touch this value if you're just adding a new option
-CONF_VERSION = '4.0.0'
+CONF_VERSION = '5.0.0'
 
 # Main configuration instance
 try: